requirements.txt
langchain
langchain-core
langchain-community
langchain-ollama
faiss-cpu
sentence-transformers
google-generativeai
python-dotenv
fastapi 
uvicorn 
pydantic 
python-multipart
numpy
scikit-learn
pandas
rouge-score
requests
